backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
SQLAlchemy==2.0.35
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
anyio==4.6.2
python-multipart>=0.0.9
